Code Formatting and Documentation

+++ modules/friends.install.php
@@ -22,24 +22,23 @@
 * @copyright  Copyright (c) 2011 Full Ambit Media, LLC (http://www.fullambit.com)
 * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
 */
-function friends_settings($data) {
+function friends_settings() {
 	return array(
-		'name' => 'friends',
+		'name'      => 'friends',
 		'shortName' => 'friends'
 	);
 }
 function friends_install($data,$drop=false) {	
 	$structures=array(
 		'friends' => array(
-			'id'										 => SQR_IDKey,
-			'user1'									 => SQR_ID,
-			'user2'									 => SQR_ID,
-			'confirmed'							 => SQR_boolean
+			'id'           => SQR_IDKey,
+			'user1'        => SQR_ID,
+			'user2'        => SQR_ID,
+			'confirmed'    => SQR_boolean
 		)
 	);
-	if($drop)
-		$data->dropTable('friends');
-	$data->createTable('friends',$structures['friends'],true);
+	if($drop) $data->dropTable('friends');
+	$data->createTable('friends',$structures['friends'],false);
 }
 function friends_uninstall($data) {
 	$data->dropTable('friends');
